Updated rayColor(), added debug Timer, minor fixes
rayColor() now creates a gradient based on the Y coordinate of the normalized ray directions.
Added a small debug Timer class that starts on construction and ends on destruction logging elapsed time.
Added constructor to Color class that takes in Vec3, to account for Vec3 arithmetic operator overloads returning a Vec3, might be changed in the future.
Minor fix to Vec3.h