Added a sphereCollision function, minor fixes
Added a function that checks if a ray collides with a given sphere, which is then used by rayColor() to decide how to color the sphere.
Moved get methods to header file in Ray class to make them inline (a fix should be found).
Minor fixes to Ray and Vec3 classes.