Added face calculation at geometry intersection
The HitRecord struct has been expanded to include a frontFace boolean that indicates whether the hit was on the front, or back face of the geometry, and the normal member will now always point in the opposite direction of the intersecting ray.
It now also includes constructors, one of which calculates the values of both frontFace and normal members.
Sphere class hit method was adapted accordingly.