Added Color::writeColor()
Added writeColor methods to the Color class, that output the color values translated to [0, 255] to an output stream.
One overload takes in account the number of samples taken neighboring regions and scales the color values back down accordingly, and then clamps them to [0, 1] before output.
Minor changes to operator<< overload for Color class.