Improved Camera class
Camera class now has a constructor that takes in the image width, height and the samplesPerPixel values.
Added an initialize method that calculates the values of the other memeber variables based on the user input from the constructors.
From now on the first pixelSample will always be from the center.