Added rtx::Material and adapted code
Added the rtx::Material abstract class, that implements the way in which a ray should scatter upon intersection with an object of said material.
Adapted rtx::Hittable::HitRecord to include a pointer to the material instance of the intersected object.
Adapted rtx::Sphere to have a material pointer member and to return it upon intersection.