Finished rtx::Dielectric implementation
Finished implementing the rtx::Dielectric material, which now checks for total internal reflections, and takes Fresnel's factor into account by using Schlick's approximation.
Whether to refract or reflect, if both are possible, is randomly decided and only one ray is scattered in one of the two directions.
Changes to the main scene.