Multithreaded scanlines, minor fixes
Scanlines (pixel rows) are now processed in parallel with the usage of the std::for_each function alongside the std::execution::par policy, which allows for automatic multithreading management.
Still retaining old code for compatibility (can be brought back by changing the value of a macro).
Made the generator for the randomFloat functions thread_local for better performance accross threads (no need for them to wait to get access to it among themselves).